+++ packages/preview/web/src/Components/default-transformation.js
@@ -2,7 +2,9 @@ import ContentStateTransformation from '../RuleEngine/ContentStateTransformation
 
 export const defaultTransformation = new ContentStateTransformation({
   _if: metadata => metadata.allText.length > 0,
-  _then: (metadata, preview) => preview.plain(metadata.allText[0]).readMore({ lines: 3 }),
+  _then: (metadata, preview) => {
+    return preview.plain(metadata.textFragments[0]).readMore({ lines: 3 });
+  },
 })
   .rule({
     _if: metadata => metadata.images.length > 0 && metadata.images.length < 5,