+++ packages/preview/web/src/ContentStateAnalyzer/ContentStateMetadata.js
@@ -1,5 +1,6 @@
 import extractEntityData from './extractEntityData';
 import { METHOD_BLOCK_MAP, METHOD_GROUPED_BLOCK_MAP } from '../const';
+import { merge, cloneDeep, groupBy } from 'lodash';
 
 const extractTextBlocksWithEntities = (blocks, entityMap, blockFilter) =>
   blocks.filter(blockFilter).reduce((texts, block) => {
@@ -23,6 +24,54 @@ const extractTextBlockArray = ({ blocks, entityMap }, blockTypeFilter) =>
     entityMap,
     ({ type, text }) => blockTypeFilter(type) && text.length > 0
   );
+
+const extractBatchesByType = ({ blocks, entityMap }, blockTypeFilter) => {
+  let current = 0,
+    next = 0;
+  const batches = groupBy(blocks, block => {
+    const isValid = blockTypeFilter(block.type);
+    if (isValid) {
+      if (current === next) next++;
+    } else if (current < next) current++;
+    return isValid && current;
+  });
+
+  const batchesWithEntities = Object.entries(batches)
+    .filter(value => value[0] !== 'false')
+    .map(batch =>
+      extractTextBlocksWithEntities(
+        batch[1],
+        entityMap,
+        ({ type, text }) => blockTypeFilter(type) && text.length > 0
+      )
+    )
+    .filter(batch => batch.length > 0);
+  return batchesWithEntities;
+};
+
+const createTextFragments = raw =>
+  extractBatchesByType(raw, type => type !== 'atomic').map(batch => {
+    if (!batch.length || batch.length === 0) return [];
+    const textCombined = batch.map(entry => entry.block.text).join(' \n');
+    const copyBlocks = cloneDeep(batch);
+    let offset = 0;
+    copyBlocks.forEach(entry => {
+      entry.block.inlineStyleRanges.map(style => (style.offset += offset));
+      entry.block.entityRanges.map(entity => (entity.offset += offset));
+      offset += entry.block.text.length + 1;
+    });
+    const inlineStyleRanges = copyBlocks.flatMap(entry => entry.block.inlineStyleRanges);
+    const entityRanges = copyBlocks.flatMap(entry => entry.block.entityRanges);
+
+    const entities = merge(copyBlocks.map(block => block.entities)).reduce((acc, curr) => ({
+      ...acc,
+      ...curr,
+    }));
+    return merge(cloneDeep(batch[0]), {
+      block: { text: textCombined, inlineStyleRanges, entityRanges },
+      entities,
+    });
+  });
 
 // extracts an array of same-type sequential block text arrays:
 // [ {li1}, {li2}, {plain}, {quote}, {li1}, {li2}, {li3} ] =>
@@ -55,7 +104,10 @@ const extractMedia = ({ entityMap }) =>
   Object.values(entityMap).reduce((media, entity) => [...media, ...extractEntityData(entity)], []);
 
 const getContentStateMetadata = raw => {
-  const metadata = { allText: extractTextBlockArray(raw, type => type !== 'atomic') };
+  const metadata = {
+    allText: extractTextBlockArray(raw, type => type !== 'atomic'),
+    textFragments: createTextFragments(raw),
+  };
 
   // non-grouped block text API
   Object.entries(METHOD_BLOCK_MAP).forEach(([func, blockType]) => {